pyproject.toml
[project]
name = "fb-oauth"
version = "0.1.0"
description = "Simple Facebook OAuth application"
readme = "README.md"
authors = [
    { name = "song", email = "slovx2@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.scripts]
fb-oauth = "fb_oauth:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
